fix: refactor dishes page files
feat: add DishesDrawer component
fix: refactor DishesNavbar and DishesLayout components
