--- src/components/Navbar.jsx
@@ -1,184 +1,0 @@
-import { useState } from 'react';
-import Link from 'next/link';
-import { useRouter } from 'next/router';
-
-import { Layout, theme, Typography, Button, Drawer, Menu } from "antd"
-const { Title } = Typography;
-const { Header } = Layout;
-import { IconChefHat, IconMoonFilled, IconSunFilled, IconMenu2 } from '@tabler/icons-react';
-import { HomeOutlined, ReadOutlined } from '@ant-design/icons';
-
-import { menuOptions } from '@/utils/menuOptions';
-
-import { useCartStore } from '@/store/cartStore';
-
-import { darkTheme } from '../../themes/dark-theme';
-
-import styles from '../styles/navbar.module.css'
-
-export const NavbarComponent = ({ themeSelected, setThemeSelected }) => {
-  const [ isDark, setIsDark] = useState(false)
-  const [open, setOpen] = useState(false);
-
-  const router = useRouter();
-
-  const cart = useCartStore((state) => state.cart)
-  const dishes = useCartStore((state) => state.dishes)
-  // const increaseCart = useCartStore((state) => state.increaseCart)
-  // const decreaseCart = useCartStore((state) => state.decreaseCart)
-  // const resetCart = useCartStore((state) => state.resetCart)
-
-  const {
-    token: { colorBgContainer },
-  } = theme.useToken();
-
-  const { useToken } = theme;
-  const { token } = useToken();
-
-  const handleDarkMode = () => {
-    if(!isDark) {
-      setThemeSelected("magentaTheme"); 
-      localStorage.setItem('theme', 'magentaTheme')
-      
-    } else {
-      setThemeSelected("darkTheme"); 
-      localStorage.setItem('theme', 'darkTheme')
-    }
-    setIsDark(!isDark)
-  }
-
-  const showDrawer = () => {
-    setOpen(true);
-  };
-
-  const onClose = () => {
-    setOpen(false);
-  };
-
-  const handleCloseDrawer = () => {
-    setTimeout(() => {
-      setOpen(false);
-    }, 350);
-  }
-  
-  return (
-    <Header 
-      className={styles.headerContainer}
-      style={{background: colorBgContainer}}
-    >
-      <nav className={styles.navContainer}>
-        <div className={styles.navContainerOptions} >
-          <Link href="/" style={{display: 'flex', alignItems: 'center'}}>
-            <IconChefHat size={36} color={`${token.colorPrimary}`}/>      
-            <Title level={4}>Magenta kitchen</Title>
-          </Link>
-
-          <Button 
-            className={styles.menuIcon} 
-            type="text" 
-            onClick={showDrawer} 
-            icon={<IconMenu2 size={30} color={token.colorPrimary} />} 
-          />
-        </div>
-
-        <div className={styles.headerButtonsContainer}>
-          <Button 
-            icon={<HomeOutlined />}
-            size='large'
-            type='text'
-            onClick={() => router.push('/')}
-          >
-            Home
-          </Button>
-
-          <Button 
-            icon={<ReadOutlined/>}
-            size='large'
-            type='text'
-            onClick={() => router.push('/dishes')}
-            // onClick={() => window.scroll({
-            //   top: 0, 
-            //   left: 0, 
-            //   behavior: 'smooth'
-            // })}
-          >
-            Menu
-          </Button>
-
-          <Button 
-            type='primary' 
-            size='large'
-            // style={{margin: '0px 20px'}}
-            onClick={() => router.push('/admin')}
-          >
-            Admin
-          </Button>
-            
-          {/* <div style={{display: 'flex',alignItems: 'center', marginRight: '20px'}}>
-            <Button 
-              icon={ isDark ? <IconSunFilled /> :<IconMoonFilled />} 
-              size='large'
-              type='text' 
-              shape='circle' 
-              onClick={handleDarkMode}
-            />
-          </div> */}
-        </div>
-      </nav>
-
-      <Drawer
-        placement="left"
-        onClose={onClose}
-        open={open}
-        width={278}
-        >
-          <div style={{background: token.colorBgContainer, paddingTop: '10px', borderRadius: '10px'}}>
-            <div  style={{ textAlign: 'center'}}>
-              <Title 
-                level={2} 
-                style={{color: token.colorPrimary,margin: '0px'}}
-              >
-                Menu 
-              </Title>
-            </div>
-            
-            <Menu
-              className={styles.drawerContainer}
-              // style={{display: 'flex', flexDirection: 'column' ,justifyContent: 'center'}}
-              // style={{background: token.colorBgElevated}}
-              // style={{width: '200px', margin: '0 auto', background: token.colorBgElevated}}
-              mode="inline"
-              defaultSelectedKeys={['1']}
-              items= {...menuOptions}
-              onClick={handleCloseDrawer}
-            />
-
-          </div>
-
-        <div>
-          <Button 
-            icon={<HomeOutlined />}
-            onClick={() => router.push('/')}
-          />
-
-          <Button 
-            type='primary' 
-            style={{margin: '0px 20px'}}
-            onClick={() => router.push('/admin')}
-          >
-            Admin
-          </Button>
-            
-          {/* <div style={{display: 'flex',alignItems: 'center', marginRight: '20px'}}>
-            <Button 
-              icon={ isDark ? <IconSunFilled /> :<IconMoonFilled />} 
-              type='text' 
-              shape='circle' 
-              onClick={handleDarkMode}
-            />
-          </div> */}
-        </div>
-      </Drawer>
-    </Header>
-  )
-}
