+++ src/components/ui/index.js
@@ -1,0 +1,3 @@
+export * from "./DishesNavbar";
+export * from "./Sidebar";
+export * from "./DishesDrawer";
