+++ src/styles/dishesNavbar.module.css
@@ -1,0 +1,5 @@
+@media (min-width: 1024px) {
+  .hamburgerMenuIcon {
+    display: none;
+  }
+}
